requirements.txt
 #============================================
# CORE AI/ML DEPENDENCIES
# ============================================

# OpenAI API and Models
openai>=1.3.0              # OpenAI GPT-4 Vision, GPT-4o models
langchain>=0.3.12          # LangChain orchestration framework
langchain-openai>=0.2.0    # LangChain OpenAI integration
langchain-core>=0.3.0      # Core LangChain utilities

# ============================================
# COMPUTER VISION DEPENDENCIES
# ============================================

# Image Processing
pillow==11.0.0             # Image manipulation and preprocessing
torchvision==0.20.1        # Computer vision models (ResNet50)
torch==2.5.1               # Deep learning framework for image encoding
transformers==4.46.3       # Hugging Face models and utilities
scikit-learn==1.5.2        # Cosine similarity and ML utilities

# ============================================
# WEB & API DEPENDENCIES
# ============================================

# Web Framework
gradio==5.22.0             # Interactive web interface for the application
requests==2.32.0           # HTTP requests (downloading images from URLs)

# Image utilities
image==1.5.33              # Additional image handling capabilities

# ============================================
# INTERACTIVE & DATA DEPENDENCIES
# ============================================

# Data Processing
pandas>=2.0.0              # Data manipulation and CSV handling
numpy>=1.24.0              # Numerical computing and array operations

# Jupyter & Interactive (optional but useful)
ipywidgets==8.1.3          # Interactive widgets for Jupyter notebooks
jupyter>=1.0.0             # Jupyter notebook support

# ============================================
# UTILITIES
# ============================================

# Environment and Configuration
python-dotenv>=1.0.0       # Load environment variables from .env file

# Google Search (optional - for additional data sources)
google-search-results==2.4.2  # Search results integration (if needed)

# ============================================
# DEVELOPMENT & DEBUGGING (OPTIONAL)
# ============================================

# Logging and Monitoring
python-json-logger>=2.0.0  # JSON logging for better debugging

# API Documentation
pydantic>=2.0.0            # Data validation and settings management

# ============================================
# NOTES
# ============================================
# 
# Key Changes from Original:
# ─────────────────────────
# 1. ✅ Removed: ibm-watsonx-ai==1.1.20
#    Replaced with: openai>=1.3.0 + langchain-openai>=0.2.0
#
# 2. ✅ Added: LangChain ecosystem for better orchestration
#    - langchain>=0.3.12
#    - langchain-core>=0.3.0
#
# 3. ✅ Added: Environment variable management
#    - python-dotenv>=1.0.0 (for storing OPENAI_API_KEY)
#
# 4. ✅ Kept: All computer vision components
#    - torch, torchvision (for ResNet50 encoding)
#    - transformers, scikit-learn (similarity matching)
#
# 5. ✅ Kept: Web interface and data processing
#    - gradio (Gradio UI)
#    - pandas, numpy (data manipulation)
#
# ============================================
# INSTALLATION INSTRUCTIONS
# ============================================
#
# 1. Create virtual environment:
#    python3.11 -m venv venv
#    source venv/bin/activate
#
# 2. Install all dependencies:
#    pip install -r requirements.txt
#
# 3. Create .env file with your API key:
#    echo "OPENAI_API_KEY=your-key-here" > .env
#
# 4. Run the application:
#    python app.py